to float

+++ pesq/pesqmain.h
@@ -1,6 +1,6 @@
 #ifndef __PESQ_MAIN_H__
 #define __PESQ_MAIN_H__
 
-double pesq(char* reference, char* degraded, int rate, char* mode);
+float pesq(char* reference, char* degraded, int rate, char* mode);
 
 #endif  // __PESQ_MAIN_H__
